Corrects the HTML tags to be applied equal spacing
In the case of the rounded box, <ul> with the class 'bk-category' is
applied equal specing and clear fix. But items for equal spacing are
<li> srounded by <ul> with 'bk-item'. Then <ul> with 'bk-item' is
the correct tag to be applied equal spacing.

One problem is that equal_spacing plugin applies equal spacing and
clear fix to the same tag.

So this pulls out the code to apply clear fix from the plugin.
It is executed in each bkviewer formatter.

And this also makes bkviewer_rounded_box to apply equal spacing to
the correct tags.