Use text() instead of html() to insert css styles
A reviewer says
	3) Please get rid of the last instances of .html(...) in favor of
	   .text(...). While the fragments passed to that function are
	   currently defined within your add-on, this might change
	   in the future, then potentially introducing security issues and
	   makes reviewing your add-on more cumbersome.

.text() cannot directly insert '<!-- -->' since it replaces '<' with '&lt'.
And CSS selector may include '>' in future. Also it should not be sanitized.
That is why I used .html().

Now I've changed my mind that this is Firefox add-on supporting
version >= 8. So it is unnecessary to comment out the CSS codes.
And the current CSS json in bkviewer_stylist_*.js does not include
CSS selector with '>'.

So I will be back to this issue when I want to use '>'.