Don't dynamically insert <script>
According to a reviewer,
	1) I get the following error after opening the page the first time and
	   the page does not show any items:
		Error: $(".bkviewer").bkviewer is not a function
		Source File: about:bkviewer
		Line: 43
	   After opening the configuration and going back, the page works

Loading configurations is perfect. The default values are used even if
there is no configuration.

The real problem is that Firefox 9.0.1 is sometimes buggy when dinamically
a <script> is inserted. According to firebug, $.fn.bkviewer() in the script
exists as a function, but it cannot be executed.

This patch statically writes all script tags, and changes the function name
bkviewer() in bkviewer_{index,rounded_box}.js to bkviewer_{index,rounded_box}()
to avoid the name conflict.

Note that I could reproduced this problem on only Windows 7 Pro 32bit/64bit,
and the configurations existed at that time.
but I've never met it on Mac OS 10.7.2 while developping bkviewer.
I also tried Ubuntu 11.4 5 times, but the problem was not reproduced.