[rounded_box]: Resize the column width on resing window size
If a title is too long and it is wraped to multiple lines
when the viewer is opened, the width property is set to the smaller size
than the width in the case that the title is written in a line.

So the column width should be resized when the window is resized.