[roundex_box]: Fixes the space under the last row in a category box
This improves the space of the left bottom corner in a category.
This assumes that font size is 16pt and 1pt=1px. The height of <li> is 25px.
The string is vertically centerized, then the space at top/bottom is
(25px - 16px) / 2 = 4.5px. The left space of the first item is 10px since
the corner radius of a category box is 10px. Then 10px - 4.5px = 5.5px is
required to be inserted under the last row.

I don't want to use 0.5px because it is rounded to 1px when it is drawed.
So this inserts 2px to top/bottom of items, and 3px as top/bottom padding of
category box.